- Created a public folder and moved public component (not admin) into it - Hotfixes for Link in Navbar.js component
