--- src/pages/Admin/User/User.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const User = () => {
-    return (
-        <div className="User">
-Liste User
-        </div>
-    );
-};
-
-export default User;
