+++ src/pages/Admin/User/Users.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Users = () => {
+    return (
+        <div className="Users">
+Liste des utilisateurs
+        </div>
+    );
+};
+
+export default Users;
