+++ src/pages/Admin/User/index.js
@@ -1,4 +1,4 @@
-export {default as User } from './User'
+export {default as Users } from './Users'
 export {default as UserAdd } from './UserAdd'
 export {default as UserEdit } from './UserEdit'
 export {default as UserDelete } from './UserDelete'
