+++ src/pages/Public/Layout.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/Public/Navbar";
 import {Outlet} from "react-router-dom";
 
 const Layout = () => {
